MedicalFormExtractor/requirements.txt
pandas
cloudpathlib
textract-trp
smart_open
boto3
requests
pymupdf
lxml
psycopg2-binary
SQLAlchemy
python-dotenv
rapidfuzz
pyyaml
pydantic